tools/requirements.txt
rdflib>=7.1.1,<8.0
pyshacl>=0.31.0
openai>=1.0
pyyaml>=6.0
